Fix Exif extraction logic when using ImageMagick
This PR removes the mapping of the file's creation date to the Exif::CREATION_DATE.
This is done in order to align it to the logic of the other two tools used for Exif extraction: Exiftool & Native.
None of this two will fill in the Exif::CREATION_DATE with the file's creation time.

This is part of the #1729 enhancement for which we need to be able to figure out if the file has a creation time as exif field or not. We can't do this currently with Lychee & ImageMagick.